Se agrega el metodo "update"  tomando id y categoria como parametros, se cambió la estructura de CategoryResponse y se agregó el metodo "delete"

+++ back/src/main/java/com/uade/back/controller/ProductController.java
@@ -37,12 +37,12 @@ public class ProductController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ProductResponse> update(@PathVariable Long id, @RequestBody ProductRequest request) {
+    public ResponseEntity<ProductResponse> update(@PathVariable Integer id, @RequestBody ProductRequest request) {
     return ResponseEntity.ok(service.update(id, request));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         service.delete(id);
     return ResponseEntity.noContent().build();
     }
